ezx-pcap: remove definitions for custom led gpio encoding
We had used these, in the leds-pcap driver, to encode gpio enabling and gpio
inversion for a led inside the variable used for the gpio number. We now rely
on gpio_is_valid() to derive if a led is gpio enabled and we have a dedicated
flag to tell if the gpio value has to be inverted.